Textures for models.

+++ Demo/common/include/GLMath.hpp
@@ -17,7 +17,7 @@
 
 typedef glm::vec4 Vec4;
 typedef glm::vec3 Vec3;
-typedef glm::ivec2 Vec2;
+typedef glm::vec2 Vec2;
 
 
 namespace GLMath
